+++ crate/receiver/script/semi-auto-deploy.sh
@@ -10,7 +10,6 @@ cd "${dir}../" || exit 1
 . ../../script/log.sh
 
 if [ "$1" = "-h" ] || [ "$1" = "--help" ]; then
-  # echo "Can start only from project folder"
   printf "Usage: %s [-h]" "${dir}"
   printf "Environment Variables:"
   printf "  SSH_USER\tSet the SSH destination as user\n"
@@ -62,7 +61,10 @@ remote_dir="/home/${SSH_USER}/pih-pah-deploy/${bin}/"
 SSH_DEST="${SSH_USER}@${SSH_ADDRESS}"
 
 log 'building...'
-env CARGO_TARGET_DIR="${dir}../../target" cargo build --release --bin receiver
+if ! env CARGO_TARGET_DIR="${dir}../../target" cargo build --release --bin receiver; then
+ error "build error"
+ exit 1
+fi
 
 # Ssh setup
 tmp_ssh_private="$(mktemp)"
@@ -70,8 +72,15 @@ echo "${SSH_PRIVATE_KEY}" > "${tmp_ssh_private}"
 
 # Transfer the Rust binary
 log 'some ssh magic...'
-ssh -o StrictHostKeyChecking=no -p "${SSH_PORT}" -i "${tmp_ssh_private}" "${SSH_DEST}" "mkdir -p ${remote_dir} && rm -f ${remote_dir}/${bin}" # if not exist
-scp -o StrictHostKeyChecking=no -P "${SSH_PORT}" -i "${tmp_ssh_private}" "${dir}../../../target/release/${bin}" "${SSH_DEST}:${remote_dir}"
+if ! ssh -o StrictHostKeyChecking=no -p "${SSH_PORT}" -i "${tmp_ssh_private}" "${SSH_DEST}" "mkdir -p ${remote_dir} && rm -f ${remote_dir}/${bin}"; then
+ error "ssh error"
+ exit 1
+fi
+
+if ! scp -o StrictHostKeyChecking=no -P "${SSH_PORT}" -i "${tmp_ssh_private}" "${dir}../../../target/release/${bin}" "${SSH_DEST}:${remote_dir}"; then
+ error "ssh error"
+ exit 1
+fi
 
 
 # SSH and setup service
@@ -103,5 +112,10 @@ WantedBy=multi-user.target" > ${TEMP_SERVICE}
 
   rm -f "${TEMP_SERVICE}"
 EOF
+# shellcheck disable=SC2181
+if [ $? -ne 0 ]; then
+ error "ssh error"
+ exit 1
+fi
 
 rm -f "${tmp_ssh_private}"
